implemented hooks for sveltekit
