+++ package.json
@@ -28,14 +28,16 @@
   },
   "devDependencies": {
     "@babel/types": "^7.24.7",
+    "@sveltejs/kit": "^2.5.18",
     "@types/estree": "^1.0.5",
     "changelogen": "^0.5.5",
-    "defu": "^6.1.4",
     "mkdist": "^1.5.3",
+    "node-html-parser": "^6.1.13",
     "typescript": "^5.5.3"
   },
   "dependencies": {
     "budoux": "^0.6.2",
+    "defu": "^6.1.4",
     "estree-walker": "^3.0.3",
     "magic-string-ast": "^0.6.2",
     "svelte": "^5.0.0-next.181",
